Aggressively minify build artifacts
tl;dr: Some unsafe, slow and extremely inefficient post processing shaves off an extra 10-20% bytes per page.

The optimizations are usually unsafe, but this is a simple enough site to manage. The "Astro" approach to this is insanely wasteful and inefficient, due to everything being processed many times over - and after bundle completion at that. However, the built static HTML is not accessible prior to that step.

At the very least it should be possible to decompose this and move the additional 'clean-css' post-processing step (which shaves off a few extra bytes on top of lightningcss - albeit unsafely as well) into a Vite plugin before it gets taken over by Astro and inserted, hundreds of times per build. Silly.

This also needed issue 509 to get a static override, due to the unescaped `<$1,000` returned from Mailchimp. It needs to be `&lt;$1,000` so that `html-minifier` doesn't choke on an unclosed opening tag.